refactor info output into sidecar source file
